test: Move `user` definition to wizard test utils
This moves `user` definition to `wizardTestUtils.tsx` so it can be shared with the test files.

The delay for user methods is also set to `null` by default as we're currently not testing with user typing latency. This should improve async behaviour and can be easily overriden for individual tests when needed.

Some early exits had to be safeguarded.