Wizard: remove Sources from Azure step (HMS-9092)
As part of the Launch decommissioning, Sources are getting rid of the
Launch option. Based on mocks provided from UX, we need to also get rid
of some parts that mention sources directly in our Wizard, and that will
be useless. This commit conditionally removes sources based on a feature
flag that we'll turn on once the decommissioning happens.