Wizard: hide azure sources integration (HMS-9092)
This commit conditionally hides the azure sources integration. It also
sets the default value of share method to 'manual' in order not to break
many things. This is a very temporary state, and we'll very soon after
turning the flag on, make these changes permanent and will deal with
more logic in the code.