playwright/CI: Fix slack notification condition in nightly action
The conditions previously were dependent on the playwright step only and fetched invalid steps property.
This also caused that during failure, the failed notification got skipped as this is Github Action behaviour.
Changed it to on success() and on failure(), since they are simpler and fit this use case better.