cockpit: use upstream cockpit types
Since we've made our typescript checks stricter, we can rely on the cockpit
types from upstream. This will prevent drift between the types. The one downside
to this is that we may have to keep the some of the cockpit lib files  and make
sure they match the commit hash in the makefile.