store/cockpit: delete cloud config
The image-builder cli handles the the available target clouds based on
environment variables set on the host. For example, if configs and
credentials exist in `~/.aws`, then aws image types will be surfaced as
available image types.

We will need a way to provide the image-build with the name of the
aws bucket, so we need to add this as an option in the AWS step of the
wizard.