Notifications: fix mutations with notifications hook
The description returned by `getErrorDescription` in case `err.status`
was not set was actually an object (`err as string` does not actually
turn the object into a string). So use `JSON.stringify` as the fallback.

Furthermore the fallback was used because the error that is passed to
`getErrorDescription` does not contain the `error` field. Rather
`status` and any additional fields are in the top-level object. This was
the case for both on prem and the service.