playwright: split original test into basic and cockpit tests
The basic test creates a blueprint, starts a build, and then deletes the
blueprint. It does not wait for the build to complete, it just makes
sure the build actually appears in the images table.

The basic test also makes sure that a toast notification with an error
is shown in case the compose request fails. This can only be tested on
the hosted UI as it is not possible to intercept requests routed through
cockpit.http in playwright.

The cockpit test covers the cloud upload dialogs which is specific to
the cockpit UI.