Playwright/CI: Add Boot tests PW project and fix CI timeout
This increases the global timeout of PW tests to 1h29.5m because of the boot
tests. Global timeout overwrites the individual test timeouts and caused
the Boot tests to fail.
Separate the Boot tests into a standalone PW project to use separate config and
change the testMatch to *.boot.ts extension.