playwright: add tests for registration (HMS-9432)
This commit adds detailed test for registration that tests automatic, register later and satellite registration. It tests behaviour of insights and rhc switch buttons, as well as different error states. Finally, it adds verification that the review step has correct values. This is a parametrized test that runs the whole workflow of create, edit, export, import and delete for all register options.