fix: correct repeatable build repositories view
This adds a different view for the 'Repositories' step that reflects if
repeatable build with a set snapshot date was chosen. It hides
information about repositories that might be incorrect at the chosen
date and present a snapshot date and displays correct package count with
a link to the snapshot detail.