fix: make packages search respect chosen date
This fixes an issue with the 'Packages' step where you could pick
additional packages that might not be available in the included
repositories at the chosen snapshot date, or ones that were removed.
This would be very complicated to do solely in this frontend so the
names search APIs were extended to support a date argument and search
the correct snapshots for the content when that's provided.