blueprints: default to host distro when importing in cockpit
The cockpit plugin does not support cross-distro building, so it's fine
to either assume or replace the distribution field.

This changes the import logic slightly, certain errors trying to map the
blueprint to the wizard state are deemed irrecoverable. Specifically the
invalid architecture and image type errors.

This also changes the tests slightly, which were previously expecting
certain blueprints to fail importing due to a missing host distro in the
blueprint. The invalid json test was actually valid json, and this has
been changes to invalid formatted json. Previously the valid, but mostly
empty, json resulted in a failure to map the blueprint to the wizard
state. Then it would fall back to treating it as an on premise
blueprint, which would fail because the host distro was missing.