Wizard: wait for changes triggered by selecting a policy
The buggy behavior of PolicySelector.tsx was this:
1. Show the selected policy name in the selector
2. Start async trigger to change additional packages, services, and kernel args

This made the boot tests crash, because PW is simply too fast,
and the trigger didn't finish.
This PR fixes that behavior by showing the policy name as the trigger applies changes.
But I'm conflicted - we are losing the point of the async trigger by waiting for it to finish
to show the changes, and a user would never be that fast to achieve this problem.
So for the sake of tests passing, and to make it more user friendly, I added an indicator for
the user to see that changes are being applied.
But I'm thinking that maybe letting the name change in the selector before the trigger ends
and forcing the tests to wait a bit could be a better option.