store/cockpit: simplify customization conversion
The hosted blueprint customizations are quite similar to
the on-prem customizations so we don't need to use a conversion
function. We can convert subscriptions and openscap options
separately. This simplifies things slightly. The initial change
to use on-prem customizations rather than the blueprint was
needed to expose registrations for the on-prem frontend.