Wizard: Remove root presence safeguards
From partitioning documentation:
"A root filesystem is added if one is not defined. This is identified by the mountpoint /. If an LVM volume group or btrfs volume is defined, the root filesystem is created as a logical volume or btrfs volume respectively, otherwise it will be created as a plain partition with a filesystem."

Therefore blocking the changes of mountpoint prefix or disabling the button to remove the root partition is not needed as its absence is resolved automatically.