src: Add 'auto-lvm' option to partitioning mode and re-enable tests
This adds a default 'auto-lvm' radio to the partitioning mode. As the partitioning mode can also be undefined a new checkbox was added. The checkbox allows the user to choose whether they'd like to specify partitioning mode used in the image or not.

There were also some disabled tests in the Disk playwright suite:
- test checking for correctly selected partitioning mode radios
- "Review exported BP" - on-prem export wasn't previously working, since the issue was fixed the test was re-enabled and used fixtures were updated.