ImageBuilderHeader: Hide OpenSourceBadge in Cockpit
This badge only makes sense in the service. While we could link to the
codebase of cockpit-image-builder aka image-builder-frontend instead of
showing the popover, we already have a hyperlink to our on premises docs
in the help icon right next to it.
Also, finding the source code isn't as hard from an RPM package as it is
from a website of a running service.