cockpit: Update doc links not to point to RHEL in Fedora
This change adds some utility hooks: one for normalising the
`process.env.IS_ON_PREMISE` variable and a second one for returning the
correct documentation link. Since we initialise the store with the host
distro, we can easily check what distro is being used. This allows us
to distinguish between RHEL on prem, Fedora on prem, and the service.
Note that this could be useful beyond just documentation links.

Previously, we were showing links to RHEL documentation all over the UI
in Fedora.

Co-authored-by: Gianluca Zuccarelli <gzuccare@redhat.com>